requirements.txt
chromadb
pandas==2.2.*
transformers==4.43.*
ragas==0.1.*
pypdf==4.3.*
python-dotenv==1.0.*
langchain==0.2.*
python-dotenv
fuzzywuzzy
pydantic>=2, <3
pytest
python=3.11.9
scipy
tiktoken
requests
langchain
langchain-community
langchain-openai
langchainhub
esprima
